% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_operations.R
\name{batch_update_schedule}
\alias{batch_update_schedule}
\title{Update a channel schedule}
\usage{
batch_update_schedule(ChannelId, Creates = NULL, Deletes = NULL)
}
\arguments{
\item{ChannelId}{Id of the channel whose schedule is being updated.}

\item{Creates}{Schedule actions to create in the schedule.}

\item{Deletes}{Schedule actions to delete from the schedule.}
}
\description{
Update a channel schedule
}
\examples{

}
