% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_operations.R
\name{create_channel}
\alias{create_channel}
\title{Creates a new channel}
\usage{
create_channel(Destinations = NULL, EncoderSettings = NULL,
  InputAttachments = NULL, InputSpecification = NULL, LogLevel = NULL,
  Name = NULL, RequestId = NULL, Reserved = NULL, RoleArn = NULL)
}
\arguments{
\item{Destinations}{}

\item{EncoderSettings}{}

\item{InputAttachments}{List of input attachments for channel.}

\item{InputSpecification}{Specification of input for this channel (max. bitrate, resolution, codec, etc.)}

\item{LogLevel}{The log level to write to CloudWatch Logs.}

\item{Name}{Name of channel.}

\item{RequestId}{Unique request ID to be specified. This is needed to prevent retries from
creating multiple resources.}

\item{Reserved}{Deprecated field that's only usable by whitelisted customers.}

\item{RoleArn}{An optional Amazon Resource Name (ARN) of the role to assume when running the Channel.}
}
\description{
Creates a new channel
}
\examples{

}
