% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_operations.R
\name{create_input}
\alias{create_input}
\title{Create an input}
\usage{
create_input(Destinations = NULL, InputSecurityGroups = NULL,
  Name = NULL, RequestId = NULL, Sources = NULL, Type = NULL)
}
\arguments{
\item{Destinations}{Destination settings for PUSH type inputs.}

\item{InputSecurityGroups}{A list of security groups referenced by IDs to attach to the input.}

\item{Name}{Name of the input.}

\item{RequestId}{Unique identifier of the request to ensure the request is handled
exactly once in case of retries.}

\item{Sources}{The source URLs for a PULL-type input. Every PULL type input needs
exactly two source URLs for redundancy.
Only specify sources for PULL type Inputs. Leave Destinations empty.}

\item{Type}{}
}
\description{
Create an input
}
\section{Accepted Parameters}{
\preformatted{create_input(
  Destinations = list(
    list(
      StreamName = "string"
    )
  ),
  InputSecurityGroups = list(
    "string"
  ),
  Name = "string",
  RequestId = "string",
  Sources = list(
    list(
      PasswordParam = "string",
      Url = "string",
      Username = "string"
    )
  ),
  Type = "UDP_PUSH"|"RTP_PUSH"|"RTMP_PUSH"|"RTMP_PULL"|"URL_PULL"
)
}
}

