% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_operations.R
\name{create_tags}
\alias{create_tags}
\title{Create tags for a resource}
\usage{
create_tags(ResourceArn, Tags = NULL)
}
\arguments{
\item{ResourceArn}{[required]}

\item{Tags}{}
}
\description{
Create tags for a resource
}
\section{Accepted Parameters}{
\preformatted{create_tags(
  ResourceArn = "string",
  Tags = list(
    "string"
  )
)
}
}

