% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_operations.R
\name{delete_input_security_group}
\alias{delete_input_security_group}
\title{Deletes an Input Security Group}
\usage{
delete_input_security_group(
  InputSecurityGroupId = "string"
)
}
\arguments{
\item{InputSecurityGroupId}{[required] The Input Security Group to delete}
}
\description{
Deletes an Input Security Group
}
\examples{

}
