% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_operations.R
\name{delete_reservation}
\alias{delete_reservation}
\title{Delete an expired reservation}
\usage{
delete_reservation(ReservationId)
}
\arguments{
\item{ReservationId}{Unique reservation ID, e.g. '1234567'}
}
\description{
Delete an expired reservation.
}
\examples{

}
