% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_operations.R
\name{delete_tags}
\alias{delete_tags}
\title{Removes tags for a resource}
\usage{
delete_tags(ResourceArn, TagKeys)
}
\arguments{
\item{ResourceArn}{[required]}

\item{TagKeys}{[required] An array of tag keys to delete}
}
\description{
Removes tags for a resource
}
\section{Accepted Parameters}{
\preformatted{delete_tags(
  ResourceArn = "string",
  TagKeys = list(
    "string"
  )
)
}
}

