% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_operations.R
\name{describe_input_security_group}
\alias{describe_input_security_group}
\title{Produces a summary of an Input Security Group}
\usage{
describe_input_security_group(
  InputSecurityGroupId = "string"
)
}
\arguments{
\item{InputSecurityGroupId}{[required] The id of the Input Security Group to describe}
}
\description{
Produces a summary of an Input Security Group
}
\examples{

}
