% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_operations.R
\name{describe_reservation}
\alias{describe_reservation}
\title{Get details for a reservation}
\usage{
describe_reservation(ReservationId)
}
\arguments{
\item{ReservationId}{[required] Unique reservation ID, e.g. '1234567'}
}
\description{
Get details for a reservation.
}
\section{Accepted Parameters}{
\preformatted{describe_reservation(
  ReservationId = "string"
)
}
}

