% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_operations.R
\name{describe_schedule}
\alias{describe_schedule}
\title{Get a channel schedule}
\usage{
describe_schedule(ChannelId, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{ChannelId}{Id of the channel whose schedule is being updated.}

\item{MaxResults}{}

\item{NextToken}{}
}
\description{
Get a channel schedule
}
\examples{

}
