% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_operations.R
\name{list_input_security_groups}
\alias{list_input_security_groups}
\title{Produces a list of Input Security Groups for an account}
\usage{
list_input_security_groups(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{}

\item{NextToken}{}
}
\description{
Produces a list of Input Security Groups for an account
}
\examples{

}
