% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_operations.R
\name{list_offerings}
\alias{list_offerings}
\title{List offerings available for purchase}
\usage{
list_offerings(
  ChannelConfiguration = "string",
  Codec = "string",
  MaxResults = 123,
  MaximumBitrate = "string",
  MaximumFramerate = "string",
  NextToken = "string",
  Resolution = "string",
  ResourceType = "string",
  SpecialFeature = "string",
  VideoQuality = "string"
)
}
\arguments{
\item{ChannelConfiguration}{Filter to offerings that match the configuration of an existing channel, e.g. '2345678' (a channel ID)}

\item{Codec}{Filter by codec, 'AVC', 'HEVC', 'MPEG2', or 'AUDIO'}

\item{MaxResults}{}

\item{MaximumBitrate}{Filter by bitrate, 'MAX_10_MBPS', 'MAX_20_MBPS', or 'MAX_50_MBPS'}

\item{MaximumFramerate}{Filter by framerate, 'MAX_30_FPS' or 'MAX_60_FPS'}

\item{NextToken}{}

\item{Resolution}{Filter by resolution, 'SD', 'HD', or 'UHD'}

\item{ResourceType}{Filter by resource type, 'INPUT', 'OUTPUT', or 'CHANNEL'}

\item{SpecialFeature}{Filter by special feature, 'ADVANCED_AUDIO' or 'AUDIO_NORMALIZATION'}

\item{VideoQuality}{Filter by video quality, 'STANDARD', 'ENHANCED', or 'PREMIUM'}
}
\description{
List offerings available for purchase.
}
\examples{

}
