% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_operations.R
\name{list_tags_for_resource}
\alias{list_tags_for_resource}
\title{Produces list of tags that have been created for a resource}
\usage{
list_tags_for_resource(ResourceArn)
}
\arguments{
\item{ResourceArn}{[required]}
}
\description{
Produces list of tags that have been created for a resource
}
\section{Accepted Parameters}{
\preformatted{list_tags_for_resource(
  ResourceArn = "string"
)
}
}

