% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_package.R
\docType{package}
\name{paws.medialive-package}
\alias{paws.medialive}
\alias{paws.medialive-package}
\title{paws.medialive: AWS Elemental MediaLive}
\description{
API for AWS Elemental MediaLive
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
