% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_operations.R
\name{purchase_offering}
\alias{purchase_offering}
\title{Purchase an offering and create a reservation}
\usage{
purchase_offering(OfferingId, Count = NULL, Name = NULL, RequestId = NULL)
}
\arguments{
\item{OfferingId}{Offering to purchase, e.g. '87654321'}

\item{Count}{Number of resources}

\item{Name}{Name for the new reservation}

\item{RequestId}{Unique request ID to be specified. This is needed to prevent retries from creating multiple resources.}
}
\description{
Purchase an offering and create a reservation.
}
\examples{

}
