% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_operations.R
\name{start_channel}
\alias{start_channel}
\title{Starts an existing channel}
\usage{
start_channel(ChannelId)
}
\arguments{
\item{ChannelId}{A request to start a channel}
}
\description{
Starts an existing channel
}
\examples{

}
