% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_operations.R
\name{stop_channel}
\alias{stop_channel}
\title{Stops a running channel}
\usage{
stop_channel(
  ChannelId = "string"
)
}
\arguments{
\item{ChannelId}{[required] A request to stop a running channel}
}
\description{
Stops a running channel
}
\examples{

}
