% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_operations.R
\name{update_channel}
\alias{update_channel}
\title{Updates a channel}
\usage{
update_channel(ChannelId, Destinations = NULL, EncoderSettings = NULL,
  InputAttachments = NULL, InputSpecification = NULL, LogLevel = NULL,
  Name = NULL, RoleArn = NULL)
}
\arguments{
\item{ChannelId}{channel ID}

\item{Destinations}{A list of output destinations for this channel.}

\item{EncoderSettings}{The encoder settings for this channel.}

\item{InputAttachments}{}

\item{InputSpecification}{Specification of input for this channel (max. bitrate, resolution, codec, etc.)}

\item{LogLevel}{The log level to write to CloudWatch Logs.}

\item{Name}{The name of the channel.}

\item{RoleArn}{An optional Amazon Resource Name (ARN) of the role to assume when running the Channel. If you do not specify this on an update call but the role was previously set that role will be removed.}
}
\description{
Updates a channel.
}
\examples{

}
