% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_operations.R
\name{batch_update_schedule}
\alias{batch_update_schedule}
\title{Update a channel schedule}
\usage{
batch_update_schedule(ChannelId, Creates = NULL, Deletes = NULL)
}
\arguments{
\item{ChannelId}{[required] Id of the channel whose schedule is being updated.}

\item{Creates}{Schedule actions to create in the schedule.}

\item{Deletes}{Schedule actions to delete from the schedule.}
}
\description{
Update a channel schedule
}
\section{Accepted Parameters}{
\preformatted{batch_update_schedule(
  ChannelId = "string",
  Creates = list(
    ScheduleActions = list(
      list(
        ActionName = "string",
        ScheduleActionSettings = list(
          HlsTimedMetadataSettings = list(
            Id3 = "string"
          ),
          InputSwitchSettings = list(
            InputAttachmentNameReference = "string"
          ),
          Scte35ReturnToNetworkSettings = list(
            SpliceEventId = 123
          ),
          Scte35SpliceInsertSettings = list(
            Duration = 123,
            SpliceEventId = 123
          ),
          Scte35TimeSignalSettings = list(
            Scte35Descriptors = list(
              list(
                Scte35DescriptorSettings = list(
                  SegmentationDescriptorScte35DescriptorSettings = list(
                    DeliveryRestrictions = list(
                      ArchiveAllowedFlag = "ARCHIVE_NOT_ALLOWED"|"ARCHIVE_ALLOWED",
                      DeviceRestrictions = "NONE"|"RESTRICT_GROUP0"|"RESTRICT_GROUP1"|"RESTRICT_GROUP2",
                      NoRegionalBlackoutFlag = "REGIONAL_BLACKOUT"|"NO_REGIONAL_BLACKOUT",
                      WebDeliveryAllowedFlag = "WEB_DELIVERY_NOT_ALLOWED"|"WEB_DELIVERY_ALLOWED"
                    ),
                    SegmentNum = 123,
                    SegmentationCancelIndicator = "SEGMENTATION_EVENT_NOT_CANCELED"|"SEGMENTATION_EVENT_CANCELED",
                    SegmentationDuration = 123,
                    SegmentationEventId = 123,
                    SegmentationTypeId = 123,
                    SegmentationUpid = "string",
                    SegmentationUpidType = 123,
                    SegmentsExpected = 123,
                    SubSegmentNum = 123,
                    SubSegmentsExpected = 123
                  )
                )
              )
            )
          ),
          StaticImageActivateSettings = list(
            Duration = 123,
            FadeIn = 123,
            FadeOut = 123,
            Height = 123,
            Image = list(
              PasswordParam = "string",
              Uri = "string",
              Username = "string"
            ),
            ImageX = 123,
            ImageY = 123,
            Layer = 123,
            Opacity = 123,
            Width = 123
          ),
          StaticImageDeactivateSettings = list(
            FadeOut = 123,
            Layer = 123
          )
        ),
        ScheduleActionStartSettings = list(
          FixedModeScheduleActionStartSettings = list(
            Time = "string"
          ),
          FollowModeScheduleActionStartSettings = list(
            FollowPoint = "END"|"START",
            ReferenceActionName = "string"
          )
        )
      )
    )
  ),
  Deletes = list(
    ActionNames = list(
      "string"
    )
  )
)
}
}

