% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_operations.R
\name{create_input_security_group}
\alias{create_input_security_group}
\title{Creates a Input Security Group}
\usage{
create_input_security_group(WhitelistRules = NULL)
}
\arguments{
\item{WhitelistRules}{List of IPv4 CIDR addresses to whitelist}
}
\description{
Creates a Input Security Group
}
\examples{

}
