% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_operations.R
\name{delete_channel}
\alias{delete_channel}
\title{Starts deletion of channel}
\usage{
delete_channel(ChannelId)
}
\arguments{
\item{ChannelId}{[required] Unique ID of the channel.}
}
\description{
Starts deletion of channel. The associated outputs are also deleted.
}
\section{Accepted Parameters}{
\preformatted{delete_channel(
  ChannelId = "string"
)
}
}

