% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_operations.R
\name{delete_input}
\alias{delete_input}
\title{Deletes the input end point}
\usage{
delete_input(InputId)
}
\arguments{
\item{InputId}{[required] Unique ID of the input}
}
\description{
Deletes the input end point
}
\section{Accepted Parameters}{
\preformatted{delete_input(
  InputId = "string"
)
}
}

