% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_operations.R
\name{describe_channel}
\alias{describe_channel}
\title{Gets details about a channel}
\usage{
describe_channel(
  ChannelId = "string"
)
}
\arguments{
\item{ChannelId}{[required] channel ID}
}
\description{
Gets details about a channel
}
\examples{

}
