% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_operations.R
\name{describe_input}
\alias{describe_input}
\title{Produces details about an input}
\usage{
describe_input(InputId)
}
\arguments{
\item{InputId}{Unique ID of the input}
}
\description{
Produces details about an input
}
\examples{

}
