% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_operations.R
\name{describe_offering}
\alias{describe_offering}
\title{Get details for an offering}
\usage{
describe_offering(OfferingId)
}
\arguments{
\item{OfferingId}{Unique offering ID, e.g. '87654321'}
}
\description{
Get details for an offering.
}
\examples{

}
