% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_operations.R
\name{list_inputs}
\alias{list_inputs}
\title{Produces list of inputs that have been created}
\usage{
list_inputs(MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{}

\item{NextToken}{}
}
\description{
Produces list of inputs that have been created
}
\section{Accepted Parameters}{
\preformatted{list_inputs(
  MaxResults = 123,
  NextToken = "string"
)
}
}

