% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.medialive_operations.R
\name{update_input}
\alias{update_input}
\title{Updates an input}
\usage{
update_input(Destinations = NULL, InputId, InputSecurityGroups = NULL,
  Name = NULL, Sources = NULL)
}
\arguments{
\item{Destinations}{Destination settings for PUSH type inputs.}

\item{InputId}{[required] Unique ID of the input.}

\item{InputSecurityGroups}{A list of security groups referenced by IDs to attach to the input.}

\item{Name}{Name of the input.}

\item{Sources}{The source URLs for a PULL-type input. Every PULL type input needs
exactly two source URLs for redundancy.
Only specify sources for PULL type Inputs. Leave Destinations empty.}
}
\description{
Updates an input.
}
\section{Accepted Parameters}{
\preformatted{update_input(
  Destinations = list(
    list(
      StreamName = "string"
    )
  ),
  InputId = "string",
  InputSecurityGroups = list(
    "string"
  ),
  Name = "string",
  Sources = list(
    list(
      PasswordParam = "string",
      Url = "string",
      Username = "string"
    )
  )
)
}
}

